Update so that generate is called passing the record in
This is a breaking change on NumberRandom which previously took a positional argument for the randomizer.

Passing the record allows for more dynamic generators to base some part of the uid off the record.  Ex. you may have a record that stores information on different media types.  You could have a generator that produced ids like: "AUDIO-1HA348" and "VIDEO-7FOA72".